+++ Kmarket2/src/main/resources/mappers/termsPolicy.xml
@@ -14,4 +14,23 @@
 		select * from `km_terms_policy_buyer`;
 	</select>
 	
+	<select id="selectTermsFinance" resultType="kr.co.kmarket2.vo.TermsFinanceVO">
+		select * from `km_terms_policy_finance`;
+	</select>
+	
+	<select id="selectTermsLocation" resultType="kr.co.kmarket2.vo.TermsLocationVO">
+		select * from `km_terms_policy_location`;
+	</select>
+	
+	<select id="selectTermsPrivacy" resultType="kr.co.kmarket2.vo.TermsPrivacyVO">
+		select * from `km_terms_policy_privacy`;
+	</select>
+
+	<select id="selectTermsSeller" resultType="kr.co.kmarket2.vo.TermsSellerVO">
+		select * from `km_terms_policy_seller`;
+	</select>	
+	
+	
+	
+	
 </mapper>
