+++ Kmarket2/src/main/resources/templates/company/index.html
@@ -57,14 +57,14 @@
     <div class="swiper">
         <header>
             <div>
-                <a th:href="@{/company/}">
+                <a th:href="@{/index}">
                     <img th:src="@{/img/company_header_logo1.png}" width="200" alt="로고">
                 </a>
                 <ul>
                     <li><a th:href="@{/company/introduce}">회사소개</a></li>
-                    <li><a href="./notice.html">전자공시</a></li>
-                    <li><a href="#">홍보채널</a></li>
-                    <li><a href="#">투명경영</a></li>
+                    <li><a th:href="@{/company/notice}">전자공시</a></li>
+                    <li><a th:href="@{/company/promote}">홍보채널</a></li>
+                    <li><a th:href="@{/company/manage}">투명경영</a></li>
                 </ul>
             </div>
         </header>
