+++ Kmarket2/src/main/resources/templates/company/introduce.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
-<html lang="en">
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
 <head>
     <meta charset="UTF-8">
     <title>Kmarket Inc. 상상하는 모든 것</title>
@@ -9,7 +10,7 @@
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/bxslider/4.2.12/jquery.bxslider.css">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.css"/>
     <script src="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.js"></script>
-    <link rel="stylesheet" href="./css/company.css">
+    <link rel="stylesheet" th:href="@{/css/company.css}">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
     <script src="https://cdn.jsdelivr.net/bxslider/4.2.12/jquery.bxslider.min.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min.js"></script>
@@ -35,14 +36,14 @@
     <div id="wrapper">
         <header>
             <div>
-                <a href="./index.html" style="color: rgb(255, 255, 255);">
-                    <img src="./kcy_img/company_header_logo2.png" width="180" alt="로고2">
+                <a th:href="@{/index}" style="color: rgb(255, 255, 255);">
+                    <img th:src="@{/img/company_header_logo2.png}" width="180" alt="로고2">
                 </a>
                 <ul>
-                    <li><a href="./introduce.html" style="color: rgb(255, 255, 255);">회사소개</a></li>
-                    <li><a href="./notice.html" style="color: rgb(255, 255, 255);">전자공시</a></li>
-                    <li><a href="#" style="color: rgb(255, 255, 255);">홍보채널</a></li>
-                    <li><a href="#" style="color: rgb(255, 255, 255);">투명경영</a></li>
+                    <li><a th:href="@{/company/index}" style="color: rgb(255, 255, 255);">회사소개</a></li>
+                    <li><a th:href="@{/company/notice}" style="color: rgb(255, 255, 255);">전자공시</a></li>
+                    <li><a th:href="@{/company/promote}" style="color: rgb(255, 255, 255);">홍보채널</a></li>
+                    <li><a th:href="@{/company/manage}" style="color: rgb(255, 255, 255);">투명경영</a></li>
                 </ul>
             </div>
         </header>
@@ -51,7 +52,12 @@
                 <div>
                     <h1>회사소개</h1>
                     <p>
-                        (문구제작)
+                        Kmarket은 
+						차별화된
+						서비스로						
+						쇼핑을 더욱 즐겁게,</br>
+						커머스를 보다
+						건강하게 만듭니다.
                     </p>
                 </div>
             </div>
@@ -59,17 +65,17 @@
                 <h2>INTRODUCTION</h2>
                 <p>Kmarket은 2014년 1월 1일 신세계백화점, 이마트 등 신세계그룹의 모든 상품을 온라인에서 한 눈에 보고, 한 번의 결제만으로 편하게 구매할 수 있는 통합 온라인 쇼핑 플랫폼으로 탄생했습니다. 차별화된 상품, 다양한 서비스, 편리한 혜택으로 삶의 가치를 높이는 쇼핑 경험을 제공합니다.</p>
                 <h2>BUSINESS AREAS</h2>
-                <img src="./kcy_img/introduce_img1.png" alt="IMG1">
-                <img src="./kcy_img/introduce_img2.png" alt="IMG2">
+                <img th:src="@{/img/introduce_img1.png}" alt="IMG1">
+                <img th:src="@{/img/introduce_img2.png}" alt="IMG2">
                 <h2>KMARKET ONLINE</h2>
-                <img src="./kcy_img/introduce_img3.png" alt="IMG3">
+                <img th:src="@{/img/introduce_img3.png}" alt="IMG3">
                 <h2>KMARKET OFFLINE</h2>
-                <img src="./kcy_img/introduce_img4.jpg" alt="IMG4">
+                <img th:src="@{/img/introduce_img4.jpg}" alt="IMG4">
             </section>
         </main>
         <footer>
             <div>
-                <img src="../img/footer_logo.png" alt="푸터로고">
+                <img th:src="@{/img/footer_logo.png}" alt="푸터로고">
                 <p>
                     대표자 : 홍길동 사업자등록번호 : 123-45-67890 부산광역시 강남구 테헤란로 123 통신판매업 신고번호 : 제2021-부산강남-12345호<br>
                     개인정보보호책임자 : 김춘추 FAX : 02-1234-5678 help@kmarket.com
