+++ Kmarket2/src/main/resources/templates/company/notice.html
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/bxslider/4.2.12/jquery.bxslider.css">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.css"/>
     <script src="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.js"></script>
-    <link rel="stylesheet" href="./css/company.css">
+    <link rel="stylesheet" th:href="@{/css/company.css}">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
     <script src="https://cdn.jsdelivr.net/bxslider/4.2.12/jquery.bxslider.min.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min.js"></script>
@@ -35,14 +35,14 @@
     <div id="wrapper">
         <header>
             <div>
-                <a href="./index.html" style="color: rgb(255, 255, 255);">
-                    <img src="./kcy_img/company_header_logo2.png" width="180" alt="로고2">
+                <a th:href="@{/index}" style="color: rgb(255, 255, 255);">
+                    <img th:src="@{/img/company_header_logo2.png}" width="180" alt="로고2">
                 </a>
                 <ul>
-                    <li><a href="./introduce.html" style="color: rgb(255, 255, 255);">회사소개</a></li>
-                    <li><a href="./notice.html" style="color: rgb(255, 255, 255);">전자공시</a></li>
-                    <li><a href="#" style="color: rgb(255, 255, 255);">홍보채널</a></li>
-                    <li><a href="#" style="color: rgb(255, 255, 255);">투명경영</a></li>
+              		<li><a th:href="@{/company/index}" style="color: rgb(255, 255, 255);">회사소개</a></li>
+                    <li><a th:href="@{/company/notice}" style="color: rgb(255, 255, 255);">전자공시</a></li>
+                    <li><a th:href="@{/company/promote}" style="color: rgb(255, 255, 255);">홍보채널</a></li>
+                    <li><a th:href="@{/company/manage}" style="color: rgb(255, 255, 255);">투명경영</a></li>
                 </ul>
             </div>
         </header>
@@ -51,17 +51,17 @@
                 <div>
                     <h1>전자공시</h1>
                     <p>
-                        (문구제작)
+                        신속하고 투명한 Kmarket
                     </p>
                 </div>
             </div>
             <section class="content">
-                <img src="./kcy_img/notice_sample.png">
+                <img th:src="@{/img/notice_sample.png}">
             </section>
         </main>
         <footer>
             <div>
-                <img src="../img/footer_logo.png" alt="푸터로고">
+                <img th:src="@{/img/footer_logo.png}" alt="푸터로고">
                 <p>
                     대표자 : 홍길동 사업자등록번호 : 123-45-67890 부산광역시 강남구 테헤란로 123 통신판매업 신고번호 : 제2021-부산강남-12345호<br>
                     개인정보보호책임자 : 김춘추 FAX : 02-1234-5678 help@kmarket.com
