+++ Kmarket2/src/main/resources/templates/termsPolicy/_footer.html
@@ -1,9 +1,9 @@
         <footer>
             <ul>
-                <li><a href="#">회사소개</a></li>
-                <li><a href="#">서비스이용약관</a></li>
-                <li><a href="#">개인정보처리방침</a></li>
-                <li><a href="#">전자금융거래약관</a></li>
+       			 <li><a th:href="@{/company/index}">회사소개</a></li>
+                 <li><a th:href="@{/termsPolicy/buyer}">서비스이용약관</a></li>
+                 <li><a th:href="@{/termsPolicy/privacy}">개인정보처리방침</a></li>
+                 <li><a th:href="@{/termsPolicy/finance}">전자금융거래약관</a></li>
             </ul>
             <div>
                 <p>
