+++ Kmarket2/src/main/resources/templates/termsPolicy/finance.html
@@ -12,50 +12,12 @@
                 </nav>
                 <article>
                   <h2>전자금융거래 약관</h2>
+                  <th:block th:each="finance : ${finances}">
                   <p>
-                    <span class="tit">제1조(목적)</span>
-                    <span class="content">이 약관은 주식회사 케이마켓(이하 “회사”라 함)가 운영하는 “K마켓” 인터넷 오픈마켓 사이트 (www.gmarket.co.kr, 이하 “K마켓”이라 함)와 스마트폰 등 이동통신기기를 통해 제공되는 ‘K마켓’ 모바일 애플리케이션(이하 “모바일K마켓”이라고 합니다)을 통해서 제공하는 전자상거래 관련 서비스 및 기타 서비스 (이하 “서비스”라 함)를 이용하는 자간의 권리, 의무를 확정하고 이를 이행함으로써 상호 발전을 도모하는 것을 그 목적으로 합니다.</span>
+                    <span class="tit">[[${finance.tit}]]</span>
+                    <span class="content">[[${finance.content}]]</span>
                   </p>
-                  <p>
-                    <span class="tit">제2조(약관의 명시, 효력과 개정)</span>
-                    <span class="content">
-                      1. 회사는 이 약관의 내용을 회사의 상호, 영업소 소재지, 대표자의 성명, 사업자등록번호, 연락처(전화, 팩스, 전자우편주소 등)등과 함께 회원이 확인할 수 있도록 “K마켓” 초기 서비스화면 또는 연결화면에 게시합니다.<br>
-                      2. 회사는 약관의규제에관한법률, 전자문서 및 전자거래기본법, 전자서명법, 정보통신망이용촉진및정보보호등에관한법률, 전자상거래등에서의소비자보호에관한법률, 전자금융거래법 등 관련법을 위배하지 않는 범위에서 본 약관을 개정할 수 있습니다.<br>
-                      3. 회사가 약관을 개정할 경우에는 적용일자 및 개정사유를 명시하여 현행 약관과 함께 초기화면에 그 적용일자 14일 이전부터 적용일자 전일까지 공지합니다.다만, 변경 내용이 중대하거나 구매회원에게 불리한 경우에는 그 적용일자 30일 이전부터 적용일자 전일까지 공지하고 구매회원에게 개별 통지합니다.변경된 약관은 그 적용일자 이전으로 소급하여 적용되지 아니합니다.<br>
-                    </span>
-                  </p>                  
-                  <p>
-                    <span class="tit">제3조</span>
-                    <span class="content">약관 내용 별도 제공</span>
-                  </p>
-                  <p>
-                    <span class="tit">제4조</span>
-                    <span class="content">약관 내용 별도 제공</span>
-                  </p>
-                  <p>
-                    <span class="tit">제5조</span>
-                    <span class="content">약관 내용 별도 제공</span>
-                  </p>
-                  <p>
-                    <span class="tit">제6조</span>
-                    <span class="content">약관 내용 별도 제공</span>
-                  </p>
-                  <p>
-                    <span class="tit">제7조</span>
-                    <span class="content">약관 내용 별도 제공</span>
-                  </p>
-                  <p>
-                    <span class="tit">제8조</span>
-                    <span class="content">약관 내용 별도 제공</span>
-                  </p>
-                  <p>
-                    <span class="tit">제9조</span>
-                    <span class="content">약관 내용 별도 제공</span>
-                  </p>
-                  <p>
-                    <span class="tit">제10조</span>
-                    <span class="content">약관 내용 별도 제공</span>
-                  </p>
+                  </th:block>
                 </article>
             </section>
